pyproject.toml
[tool.poetry]
name = "color-picker"
version = "0.1.0"
description = ""
authors = ["chebupelka8 <stpzamyatin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pillow = "^10.3.0"
pyside6 = "^6.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
